pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydecorrelation-stretch"
version = "0.6.0"
description = "Python implementation of DStretch decorrelation stretch for rock art analysis"
authors = [{name = "Víctor Méndez", email = "victor.mendez@uc.cl"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "opencv-python>=4.5.0",
    "pillow>=8.3.0",
    "matplotlib>=3.7.5",
    "scikit-image>=0.21.0",
    "nuitka>=2.7.16",
    "commitizen>=3.31.0",
    "pytest>=8.3.5",
    "ruff>=0.14.10",
    "pyright>=1.1.407",
    "numba>=0.58.1",
    "psutil>=7.2.1",
    "customtkinter>=5.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
]

[project.scripts]
dstretch = "pydecorrelation_stretch.cli:main"
dstretch-gui = "pydecorrelation_stretch.gui:main"

[project.urls]
Homepage = "https://github.com/arqueomendez/pydecorrelation-stretch"
Repository = "https://github.com/arqueomendez/pydecorrelation-stretch.git"

[tool.hatch.build.targets.wheel]
packages = ["src/pydecorrelation_stretch"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
    "/README.md",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = "src"
addopts = "-v"

[tool.pyright]
include = ["src"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
